requirements.txt
alembic==1.13.1
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==2.2.5
Flask-Cors==4.0.1
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
packaging==24.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==2.3.7
WTForms==3.1.2
